requirements.txt
asgiref==3.4.1
crispy-bootstrap5==0.6
Django==3.2.9
django-countries==7.2.1
django-crispy-forms==1.13.0
django-localflavor==3.1
django-phonenumber-field==6.0.0
phonenumberslite==8.12.38
python-stdnum==1.17
pytz==2021.3
sqlparse==0.4.2
coverage>=3.7.1
flake8
icalendar>=3.8.4
isort
python-dateutil>=2.1
wheel
